fix(realtime): lfilter float64 → MPS crash on speech
scipy.signal.lfilter always returns float64 regardless of float32 input.
The EQ block (infer_wav.cpu().numpy() → lfilter → from_numpy → .to(mps))
was producing a float64 tensor which MPS rejects with:
  'Cannot convert a MPS Tensor to float64 dtype'

This only triggered when speaking (gate_gain > 0 → EQ runs). The warmup
and silence path both skip the EQ block entirely — which is why the 20s
test with silence passed, but the session crashed on first real audio.

The barebone worker written during debugging happened to work because it
had no EQ at all, masking the root cause.

Fix: cast both infer_np and _lp_component to float32 explicitly before
subtraction so the result stays float32 through the MPS upload.

+++ backend/app/_realtime_worker.py
@@ -508,8 +508,8 @@ def run_worker(
                 # spectral centroid +800 Hz above input). Apply ~-3 dB shelf above
                 # ~3 kHz by subtracting 50% of the high-pass filtered signal.
                 # Coefficients pre-computed above; no cross-block IIR state needed.
-                infer_np = infer_wav.cpu().numpy()
-                _lp_component = scipy.signal.lfilter(_hs_b, _hs_a, infer_np)
+                infer_np = infer_wav.cpu().numpy().astype(np.float32)
+                _lp_component = scipy.signal.lfilter(_hs_b, _hs_a, infer_np).astype(np.float32)
                 infer_np = infer_np - 0.50 * _lp_component
                 infer_wav = _torch.from_numpy(infer_np).to(infer_device)
